requirements.txt
numpy
Pillow
Keras
tensorflow
Flask
Werkzeug
gunicorn
efficientnet
python== 3.8
tensorflow~=2.16.1
numpy~=1.26.4
matplotlib~=3.8.4
scikit-learn~=1.4.2
pillow~=10.3.0
flask~=3.0.3
keras~=3.3.3